MIR-616 Use Yarn instead of npm
- switched from npm to yarn
- moved GruntFile.js and package.json, to use standard mycore yarn build process
