+++ mir-module/package.json
@@ -1,6 +1,4 @@
 {
-    "name": "${project.artifactId}",
-    "description": "${project.description}",
     "version": "1.0.0",
     "author": "René Adler",
     "homepage": "${project.url}",
